# src/main/resources/templates/openapi.yaml
openapi: 3.0.3
info:
  title: Swagger lokapi - OpenAPI 3.0
  description: ''
  termsOfService: http://swagger.io/terms/
  contact:
    email: tawfik.elyousfi@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: https://test.swagger.io/api/v3
...
paths:
  /file/preview/{filename}:
    get:
      tags:
        - File
      summary: Preview a file directly without downloading it
      description: Streams a file as a binary resource for preview.
      operationId: previewFile
      parameters:
        - name: filename
          in: query
          required: true
          schema:
            type: string
          description: The name of the file to preview.
      responses:
        '200':
          description: The file was successfully streamed for preview.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
              example: "Binary content for preview"
        '404':
          description: File not found on the server.
        '500':
          description: Server error while processing the file.
  /file/download/{filename}:
    get:
      tags:
        - file
      summary: Download a file from the server
      operationId: downloadFile
      parameters:
        - name: filename
          in: query
          required: true
          schema:
            type: string
          description: File name
      responses:
        200:
          description: File downloaded successfully
          content:
            application/octet-stream:  # MIME type pour indiquer un fichier binaire
              schema:
                type: string
                format: binary  # Indique que la réponse est un fichier binaire
        400:
          description: Invalid parameters or file not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /file/upload:
    post:
      tags:
        - file
      summary: Upload files from application
      operationId: uploadFile
      parameters:
        - name: owner
          in: query
          required: true
          schema:
            type: string
          description: L'ID du propriétaire (owner)
        - name: tenant
          in: query
          required: true
          schema:
            type: string
          description: L'ID du locataire (tenant)
        - name: property
          in: query
          required: true
          schema:
            type: string
          description: L'adresse de la propriété (property)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Le fichier PDF à uploader.
              required:
                - file
      responses:
        default:
          description: Fichier uploadé avec succès
          content:
            application/json:
              schema:
                type: string

        '400':
          description: Aucun fichier sélectionné ou mauvaise requête
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Aucun fichier sélectionné.
        '415':
          description: Mauvais type de fichier
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Seuls les fichiers PDF sont acceptés.
        '500':
          description: Erreur interne du serveur
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Erreur lors de l'enregistrement du fichier.
  /auth/register:
    post:
      tags:
        - auth
      summary: Register a new user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: 'newUser'
                email:
                  type: string
                  example: 'newuser@email.com'
                password:
                  type: string
                  example: 'securePassword123'
                roles:
                  type: array
                  items:
                    type: string
                    enum:
                      - OWNER
                      - TENANT
                  example:
                    - TENANT
                    - OWNER
                  description: tenant or owner status in application
      responses:
        default:
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMessageResponse'
        '400':
          description: Invalid input
  /auth/infoUser/{id}:
    get:
      tags:
        - auth
      summary: Getinf from user
      operationId: infoUser
      parameters:
        - name: userId
          in: path
          description: 'The id that needs to be fetched. Use user for testing. '
          required: true
          schema:
            type: integer
            format: int64
      responses:
        default:
          description: 'User authenticated successfully returned'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: 'Get info failed'
  /auth/login:
    post:
      tags:
        - auth
      summary: Authenticate a user
      operationId: authenticateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: 'username'
                email:
                  type: string
                  example: 'email@email.com'
                password:
                  type: string
                  example: 'securePassword123'
      responses:
        default:
          description: 'User authenticated successfully token returned'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthToken'
        '400':
          description: 'Authentication failed'
  /auth/activate:
    get:
      tags:
        - auth
      summary: Activate a user registered
      operationId: activateUser
      parameters:
        - name: token
          in: path
          description: 'Token to activate the user '
          required: true
          schema:
            type: string
      #requestBody:
      #  required: true
      #  content:
      #    application/json:
      #      schema:
      #        type: object
      #        properties:
      #          token:
      #            type: string
      #            example: 'a1b2c3d4e5f6g7h8i9j0'*/
      responses:
        default:
          description: Message User actived successfully
          content:
            application/json:
              schema:
                type: string
        '400':
          description: Authentication failed
...
  /users/{userId}:
    get:
      tags:
        - user
      summary: Get user by user Id
      description: ''
      operationId: getUserById
      parameters:
        - name: userId
          in: path
          description: 'The id that needs to be fetched. Use user1 for testing. '
          required: true
          schema:
            type: integer
            format: int64
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid ID supplied
        '404':
          description: User not found
    put:
      tags:
        - user
      summary: Update user
      description: This can only be done by the logged in user.
      operationId: updateUserById
      parameters:
        - name: userId
          in: path
          description: ID that need to be updated
          required: true
          schema:
            type: integer
            format: int64
      #requestBody:
      #  description: Update an existent user in the store
      #  content:
      #    application/json:
      #      schema:
      #        $ref: '#/components/schemas/User'
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      tags:
        - user
      summary: Delete user
      description: This can only be done by the logged in user.
      operationId: deleteUserById
      parameters:
        - name: userId
          in: path
          description: The ID that needs to be deleted
          required: true
          schema:
            type: integer
            format: int64
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid ID supplied
        '404':
          description: User not found
  /users/{username}:
    get:
      tags:
        - user
      summary: Get user by user name
      description: ''
      operationId: getUserByName
      parameters:
        - name: username
          in: path
          description: 'The name that needs to be fetched. Use user1 for testing. '
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid username supplied
        '404':
          description: User not found
...
components:
  schemas:
    UserMessageResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        message:
          type: string
      required:
        - message
    AuthToken:
      type: object
      properties:
        token:
          type: string
          description: "JWT Token"
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.e..."
        message:
          type: string
          description: "Message comportement"
          example: "message d'information"
        user:
          $ref: '#/components/schemas/User'
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        username:
          type: string
          example: 'theUser'
        email:
          type: string
          example: 'you@email.com'
        password:
          type: string
          example: '12345'
        roles:
          type: array
          items:
            type: string
            enum:
              - OWNER
              - TENANT
          example:
            - TENANT
            - OWNER
          description: tenant or owner status in application

        active:
            type: boolean
            example: 1
        creationDate:
          type: string
          format: date
          description: User creation date
        updateDate:
          type: string
          format: date
          description: User update date
        deleteDate:
          type: string
          format: date
          description: User detele date
      required:
        - id
        - username
        - email
        - password
        - roles
...